Changes to reset position functionality
Changed the plotter to reset position to origo when no command is received from mDraw for a second.

Earlier the plotter position was manually reset when an "M4" command was received, but this is the command for driving the laser so the program wouldn't have been functional if the laser functionality was implemented.